requirements.txt
Flask==3.0.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core>=0.3.64
langchain-ollama==0.3.2
langchain-pinecone==0.0.1
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.26.4
openai==1.76.0
pinecone-client==3.0.2
python-dotenv>=1.0.0
requests>=2.31.0
torch==2.5.1
transformers==4.46.2
Werkzeug>=3.0.0
